requirements.txt
annotated-types==0.6.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cymem==2.0.8
decorator==4.4.2
faiss-cpu==1.7.4
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
moviepy==1.0.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pillow==10.2.0
preshed==3.0.9
proglog==0.1.10
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
srsly==2.4.8
sympy==1.12
tabulate==0.9.0
thinc==8.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
wasabi==1.1.2
weasel==0.3.4
